+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### HEAD
+* Add option to configure Browsersync ([#1822](https://github.com/roots/sage/pull/1822))
+* Add support for HTML injection ([#1817](https://github.com/roots/sage/pull/1817))
+
 ### 9.0.0-beta.2: January 19th, 2016
 * Fix Browersync ([#1815](https://github.com/roots/sage/pull/1815))
 * Add option to select CSS framework, add Foundation as an option ([#1813](https://github.com/roots/sage/pull/1813))